@@ -3,58 +3,58 @@
 	"version": "0.0.0",
 	"license": "MIT",
 	"scripts": {
-		"prestart": "npm run build && cd src/server && tsc && npm run generate-sw",
+		"build": "ng build --prod -aot --output-hashing=none && ng build --prod --app server --output-hashing=none",
+		"generate-sw": "node generate-sw.js",
+		"prestart": "npm run build && npm run tsc-server && npm run generate-sw",
 		"start": "node dist/server.js",
-		"build": "ng build --prod -aot && ng build --prod --app server --output-hashing=none",
-		"generate-sw": "node generate-sw.js"
+		"tsc-server": "cd src/server && tsc"
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animation": "^4.0.0-beta.8",
-		"@angular/animations": "^4.4.5",
-		"@angular/app-shell": "^0.1.0",
-		"@angular/benchpress": "^0.1.0",
+		"@angular/animations": "^5.0.0",
 		"@angular/cdk": "^2.0.0-beta.12",
+		"@angular/common": "^5.0.0",
+		"@angular/compiler": "^5.0.0",
+		"@angular/core": "^5.0.0",
 		"@angular/flex-layout": "^2.0.0-beta.9",
-		"@angular/common": "^4.4.5",
+		"@angular/forms": "^5.0.0",
+		"@angular/http": "^5.0.0",
 		"@angular/material": "^2.0.0-beta.12",
-		"@angular/compiler": "^4.4.5",
-		"@angular/core": "^4.4.5",
-		"@angular/forms": "^4.4.5",
-		"@angular/http": "^4.4.5",
-		"@angular/platform-browser": "^4.4.5",
-		"@angular/platform-browser-dynamic": "^4.4.5",
-		"@angular/router": "^4.4.5",
-		"@angular/platform-server": "^4.4.5",
-		"@angular/service-worker": "^1.0.0-beta.16",
+		"@angular/platform-browser": "^5.0.0",
+		"@angular/platform-browser-dynamic": "^5.0.0",
+		"@angular/platform-server": "^5.0.0",
+		"@angular/platform-webworker": "^5.0.0",
+		"@angular/platform-webworker-dynamic": "^5.0.0",
+		"@angular/router": "^5.0.0",
+		"@angular/service-worker": "^5.0.0",
 		"@nguniversal/common": "^5.0.0-beta.1",
 		"@nguniversal/express-engine": "^1.0.0-beta.3",
 		"@nguniversal/module-map-ngfactory-loader": "^1.0.0-beta.3",
-		"@angular/platform-webworker": "^4.4.5",
-		"@angular/platform-webworker-dynamic": "^4.4.5",
 		"animate.css": "^3.5.2",
 		"classlist.js": "^1.1.20150312",
+		"core-js": "^2.5.1",
 		"font-awesome": "^4.7.0",
 		"hammerjs": "^2.0.8",
 		"idb-keyval": "latest",
 		"intl": "^1.2.5",
 		"materialize-css": "^0.100.2",
 		"moment": "^2.18.1",
-		"core-js": "^2.5.1",
-		"rxjs": "^5.4.3",
-		"socket.io": "^2.0.3",
+		"rxjs": "^5.5.2",
+		"socket.io": "^2.0.4",
 		"socketio-file-upload": "^0.6.0",
 		"sse": "0.0.6",
+		"ejs": "latest",
+		"ts-loader": "^3.1.1",
 		"web-animations-js": "^2.3.1",
 		"workbox-build": "latest",
 		"workbox-sw": "latest",
-		"zone.js": "^0.8.14"
+		"zone.js": "^0.8.18"
 	},
 	"devDependencies": {
-		"@angular/cli": "1.5.0-rc.0",
-		"@angular/compiler-cli": "^4.4.5",
-		"@angular/language-service": "^4.4.5",
-		"typescript": "~2.5.3",
-		"webpack": "^3.7.1"
+		"@angular/cli": "^1.5.0",
+		"@angular/compiler-cli": "^5.0.0",
+		"@angular/language-service": "^5.0.0",
+		"typescript": "~2.6.0",
+		"webpack": "^3.8.1"
 	}
 }
