@@ -1,4 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
+
+import { _ContainerComponent } from '../../modules/_dry/index';
+import { NavComponent_ } from '../_nav/_nav.component';
 
 @Component({
 	selector: 'pp2-home',
@@ -6,7 +9,13 @@ import { Component, OnInit } from '@angular/core';
 })
 
 export class HomeComponent implements OnInit {
-	constructor() { }
-
-	ngOnInit() { }
+	@ViewChild('C_Pp2_Dry_Container') C_Pp2_Dry_Container: _ContainerComponent;
+	@ViewChild('C_Pp2__Nav') C_Pp2__Nav: NavComponent_;
+	constructor() {}
+	ngAfterViewInit(){
+		this.C_Pp2__Nav.$C_Mat_Sidenav_Click$.subscribe(() => {
+			this.C_Pp2_Dry_Container.C_Mat_Sidenav.toggle();
+		})
+	}
+	ngOnInit(){}
 }
