@@ -1,25 +1,35 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
+import { MatSidenav } from '@angular/material';
+
 import { Pp2MediaQueryService } from '../../services/Pp2-media-query.service';
-import { Pp2MediaQueryScreenConst } from '../../consts/Pp2MediaQueryScreen.const';
-import { Pp2MediaQueryScreen } from '../../interfaces/Pp2-media-query.interface';
 
 @Component({
 	selector: 'pp2-dry-container',
 	templateUrl: '_container.component.html',
-	styles: [``]
+	styles: [`
+		.mat-sidenav-content{
+			height: calc(100vh - 64px);
+			overflow-y: auto;
+			position: fixed;
+			top:64px;
+			width:100vw;
+		}
+		@media (max-width: 600px) {
+			.mat-sidenav-content{
+				height: calc(100vh - 56px);
+				top:56px;
+			}
+		}
+	`]
 })
-
 export class _ContainerComponent implements OnInit {
-	$$_Pp2MQ: { screen: Pp2MediaQueryScreen } = { screen: Pp2MediaQueryScreenConst };
+	@ViewChild('C_Mat_Sidenav') C_Mat_Sidenav: MatSidenav;
 	constructor(
 		public $_Pp2MQ: Pp2MediaQueryService
-	) {
-		this.$_Pp2MQ.screen.subscribe((screen: Pp2MediaQueryScreen) => {
-			this.$$_Pp2MQ.screen = screen;
-		});
+	) {}
+	ngOnDestroy() {}
+	ngOnInit() {}
+	gg() {
+		return this.$_Pp2MQ.screen.lt.xl ? 'FOO' : 'BAR';
 	}
-	ngOnDestroy() {
-		// this.$_Pp2MQ.screen.unsubscribe();
-	}
-	ngOnInit() { }
 }
