@@ -1,11 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit } from "@angular/core";
 
 @Component({
-	selector: 'pp2-su-root',
-	templateUrl: 'su.component.html'
+  selector: 'pp2-su-root',
+  template: '<router-outlet></router-outlet>'
 })
-
 export class Pp2SuComponent implements OnInit {
-	constructor() { }
-	ngOnInit() { }
+  constructor() {}
+  ngOnInit() {}
 }
