@@ -9,14 +9,11 @@ declare var window: Window;
 })
 export class Pp2Component implements OnInit {
 	title = 'pp2';
-	private isSubscribed = false;
-	private registration = undefined;
 	ngOnInit() {
-		if (!isDevMode()) {
-			const swRegister = window.document.createElement('script');
-			swRegister.src = 'assets/sw-register.js';
-			window.document.body.appendChild(swRegister);
-			console.log('!isDevMode')
-		}
+		// if (!isDevMode()) {
+		// 	const swRegister = window.document.createElement('script');
+		// 	swRegister.src = 'assets/sw-register.js';
+		// 	window.document.body.appendChild(swRegister);
+		// }
 	}
 }
