@@ -1,30 +1,27 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+// import { TransferHttpCacheModule } from '@nguniversal/common';
 
 import { DryModule } from './modules/_dry/dry.module';
 
-import { Pp2RoutingModule } from './pp2.routing.module';
+import { Pp2RoutingModule, Pp2Components } from './pp2.routing.module';
 
 import { Pp2SuModule } from './modules/su/su.module';
 
 import { Pp2Component } from './pp2.component';
-import { HomeComponent } from './components/home/home.component';
-import { DevComponent } from './components/dev/dev.component';
-
-export const Pp2Components = [
-	Pp2Component,
-	HomeComponent,
-	DevComponent
-];
+import { NavComponent_ } from './components/_nav/_nav.component';
 
 @NgModule({
-	declarations: [
-		...Pp2Components
-	],
 	imports: [
 		BrowserModule.withServerTransition({ appId: 'pp2' }),
-		Pp2RoutingModule,
-		Pp2SuModule
+		// TransferHttpCacheModule,
+		Pp2SuModule,
+		Pp2RoutingModule
+	],
+	declarations: [
+		Pp2Component,
+		NavComponent_,
+		...Pp2Components
 	],
 	bootstrap: [Pp2Component]
 })
